requirements.txt
pbr>=1.10.0,<5.8.1  # Apache-2.0
oslo.config!=3.18.0,>=3.22.2,<10.0.0  # Apache-2.0
setuptools>=75.3.0,<76.0.0  # MIT License
PyYAML>=6.0.2,<7.0.0  # MIT License
GitPython>=3.1.44,<4.0.0  # BSD License (BSD-3-Clause)
Jinja2>=3.1.5,<4.0.0
